Improve zshrc
Make the `gitstatus` plugin display a yellow section if the current,
local branch is not up to date with the remote branch, ie. it is a few
commits behind/ahead of it.

Manually add LS_COLORS and PAGER variables to .zshrc. Also, manually
export locale variables.

In prompt, when tab completing, highlight current item.